+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 ### Next
 
+- Add support for custom encoder models. See https://github.com/flekschas/peax-avocado for an example
+- Add support for custom data directory via the `--base-data-dir` argument of `start.py` 
+
+### v0.4.0
+
+This is the version described in our paper (Lekschas et al., 2020).
+
 - Add support for changing the classifier via the config file's `classifier` and `classifier_params` properties
 - Add ability to show more than 5 results per page via to HiGlass' new view scrolling
 - Add legend to the UMAP projection view and highlight